B-sim-floats: eliminate floating-point from deterministic sim
Purge all f32/f64 from elven_canopy_sim to guarantee bit-identical
cross-platform determinism for lockstep multiplayer and replays.

Phase A — sim-state fields:
- Task progress and total_cost converted to i64 (1000x scaled ticks)
- Tree mana stored as millimana (i64), fruit yields as PPM (i64)
- Greenhouse efficiency integer per-mille; mood/mana multipliers i64
- Config fields (mana costs, regen rates, hunger thresholds) all integer

Phase B — nav distances and pathfinding:
- Nav edge distances use scaled-integer sqrt (1024x fixed-point)
- A* open set keyed on i64 cost; movement and combat delays use div_ceil
- Melee/ranged delay, pursuit radius, flee distance all integer

Phase C — raycast, tree generation, structural solver:
- DDA line-of-sight (has_los) rewritten as pure-integer Bresenham-style
- tree_gen fully rewritten with i64 fixed-point arithmetic and a 256-entry
  sine lookup table; branch angles, tapering, and crown geometry all integer
- Structural FEM solver converted to i64 fixed-point (FP_SCALE = 1_000_000)
  with integer stiffness matrices, force vectors, and displacement solve

Once-over fixes:
- Fixed fp_harmonic_mean returning 0 for single-element input
- Added 11 new tests (task serde, nav distance symmetry, A* i64 costs,
  integer movement delay, structural FP arithmetic, tree_gen sine table,
  DDA raycast, fp_harmonic_mean, config PPM validation)
- Doc and comment cleanup across all changed modules

+++ elven_canopy_sim/src/sim/construction.rs
@@ -235,8 +235,8 @@ impl SimState {
             kind: task::TaskKind::Build { project_id },
             state: task::TaskState::Available,
             location: task_location,
-            progress: 0.0,
-            total_cost: num_voxels as f32,
+            progress: 0,
+            total_cost: num_voxels as i64,
             required_species: Some(Species::Elf),
             origin: task::TaskOrigin::PlayerDirected,
             target_creature: None,
@@ -390,8 +390,8 @@ impl SimState {
             kind: task::TaskKind::Build { project_id },
             state: task::TaskState::Available,
             location: task_location,
-            progress: 0.0,
-            total_cost: num_voxels as f32,
+            progress: 0,
+            total_cost: num_voxels as i64,
             required_species: Some(Species::Elf),
             origin: task::TaskOrigin::PlayerDirected,
             target_creature: None,
@@ -546,8 +546,8 @@ impl SimState {
             kind: task::TaskKind::Build { project_id },
             state: task::TaskState::Available,
             location: task_location,
-            progress: 0.0,
-            total_cost: num_voxels as f32,
+            progress: 0,
+            total_cost: num_voxels as i64,
             required_species: Some(Species::Elf),
             origin: task::TaskOrigin::PlayerDirected,
             target_creature: None,
@@ -668,8 +668,8 @@ impl SimState {
             kind: task::TaskKind::Build { project_id },
             state: task::TaskState::Available,
             location: task_location,
-            progress: 0.0,
-            total_cost: num_voxels as f32,
+            progress: 0,
+            total_cost: num_voxels as i64,
             required_species: Some(Species::Elf),
             origin: task::TaskOrigin::PlayerDirected,
             target_creature: None,
@@ -832,8 +832,8 @@ impl SimState {
             kind,
             state: task::TaskState::Available,
             location,
-            progress: 0.0,
-            total_cost: 0.0,
+            progress: 0,
+            total_cost: 0,
             required_species,
             origin: task::TaskOrigin::PlayerDirected,
             target_creature: None,
@@ -862,13 +862,8 @@ impl SimState {
         };
 
         // Mark composition as build_started on the first Build action.
-        let progress = self
-            .db
-            .tasks
-            .get(&task_id)
-            .map(|t| t.progress)
-            .unwrap_or(0.0);
-        if progress == 0.0
+        let progress = self.db.tasks.get(&task_id).map(|t| t.progress).unwrap_or(0);
+        if progress == 0
             && let Some(bp) = self.db.blueprints.get(&project_id)
             && let Some(comp_id) = bp.composition_id
         {
@@ -933,7 +928,7 @@ impl SimState {
 
         // Increment progress by 1 (one voxel).
         let _ = self.db.tasks.modify_unchecked(&task_id, |t| {
-            t.progress += 1.0;
+            t.progress += 1;
         });
 
         // Check if the build is complete.
@@ -1399,14 +1394,14 @@ impl SimState {
         };
 
         // Create the Furnish task. total_cost = number of furniture items.
-        let total_cost = planned_count as f32;
+        let total_cost = planned_count as i64;
         let task_id = TaskId::new(&mut self.rng);
         let new_task = task::Task {
             id: task_id,
             kind: task::TaskKind::Furnish { structure_id },
             state: task::TaskState::Available,
             location,
-            progress: 0.0,
+            progress: 0,
             total_cost,
             required_species: Some(Species::Elf),
             origin: task::TaskOrigin::PlayerDirected,
@@ -1530,7 +1525,7 @@ impl SimState {
 
         // Increment progress by 1 (one item).
         let _ = self.db.tasks.modify_unchecked(&task_id, |t| {
-            t.progress += 1.0;
+            t.progress += 1;
         });
 
         // Check if furnishing is complete.
